Add files via upload
mandelbrot.py includes a function that generates a plot of the Mandelbrot Set, with inputs for n -> size of grid, N_max -> number of iterations, and threshold -> determining whether points belong in set. The function is then tested with respective values of 5000, 50, and 50.
markov_chain.py includes a function that simulates a random vector p and random transition matrix P, with a Markov chain for an inputted number of iterations. The difference between the stationary point and the random vector p, after each iteration, is then plotted.